pyproject.toml
[project]
name = "crypto-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt>=4.5.31",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
